Fix editor initialization in relation manager modals and reordered repeaters
In the newest version of Laravel opening a relation manager modal will
not render tiny editor after the first time. This is because Laravel's
id generation code has changed in a way which causes tinymce to not
re-initialize after the first call for a given relation manager modal
because the element's id is unchanged.

Now the code listens for elements removed from the DOM and cleans up
tinymce editor instances attached to removed elements. This is also
a performance improvement because previously tiny editors associated
with removed inputs would hang around in memory.